iq-bot-writer/requirements.txt
openai
redis
python-dotenv
requests
-e ../iq-bot-global